# inflate: Keep dict on stack